+++ crates/chainio/clients/avsregistry/src/reader.rs
@@ -17,6 +17,7 @@ use eigen_types::operator::{
 use eigen_utils::slashing::middleware::blsapkregistry::BLSApkRegistry;
 use eigen_utils::slashing::middleware::operatorstateretriever::OperatorStateRetriever;
 use eigen_utils::slashing::middleware::registrycoordinator::RegistryCoordinator;
+use eigen_utils::slashing::middleware::servicemanagerbase::ServiceManagerBase;
 use eigen_utils::slashing::middleware::stakeregistry::IStakeRegistryTypes::{
     StakeUpdate, StrategyParams,
 };
@@ -321,6 +322,32 @@ impl AvsRegistryChainReader {
         Ok(s)
     }
 
+    /// Get a list of strategies that the AVS supports for restaking
+    ///
+    /// # Returns
+    ///
+    /// A vector of addresses representing the strategies that the AVS supports for restaking
+    pub async fn get_restakeable_strategies(&self) -> Result<Vec<Address>, AvsRegistryError> {
+        let provider = get_provider(&self.provider);
+
+        let contract_registry_coordinator =
+            RegistryCoordinator::new(self.registry_coordinator_addr, &provider);
+
+        let service_manager = contract_registry_coordinator
+            .serviceManager()
+            .call()
+            .await?
+            ._0;
+
+        let strategies = ServiceManagerBase::new(service_manager, &provider)
+            .getRestakeableStrategies()
+            .call()
+            .await?
+            ._0;
+
+        Ok(strategies)
+    }
+
     /// Get operators stake in quorums of operator at current block
     ///
     /// # Arguments
@@ -436,6 +463,39 @@ impl AvsRegistryChainReader {
             }
         }
         Ok(quorums)
+    }
+
+    /// Get a list of strategies that an operator has potentially restaked on the AVS.
+    ///
+    /// # Arguments
+    ///
+    /// * `operator_address` - The operator address.
+    ///
+    /// # Returns
+    ///
+    /// A vector with strategies addresses
+    pub async fn get_operator_restaked_strategies(
+        &self,
+        operator_address: Address,
+    ) -> Result<Vec<Address>, AvsRegistryError> {
+        let provider = get_provider(&self.provider);
+
+        let contract_registry_coordinator =
+            RegistryCoordinator::new(self.registry_coordinator_addr, &provider);
+
+        let service_manager = contract_registry_coordinator
+            .serviceManager()
+            .call()
+            .await?
+            ._0;
+
+        let strategies = ServiceManagerBase::new(service_manager, &provider)
+            .getOperatorRestakedStrategies(operator_address)
+            .call()
+            .await?
+            ._0;
+
+        Ok(strategies)
     }
 
     /// Get operator id
@@ -1084,33 +1144,20 @@ impl AvsRegistryChainReader {
 
 #[cfg(test)]
 mod tests {
+
     use super::*;
-    use crate::test_utils::{build_avs_registry_chain_writer, test_register_operator};
-    use eigen_logging::get_test_logger;
+    use crate::test_utils::{
+        build_avs_registry_chain_reader, build_avs_registry_chain_writer, test_register_operator,
+    };
+    use eigen_crypto_bls::BlsKeyPair;
     use eigen_testing_utils::{
         anvil::{start_anvil_container, start_m2_anvil_container},
         anvil_constants::{
-            get_operator_state_retriever_address, get_registry_coordinator_address, FIFTH_ADDRESS,
-            FIFTH_PRIVATE_KEY, FIRST_ADDRESS, OPERATOR_BLS_KEY,
+            FIFTH_ADDRESS, FIFTH_PRIVATE_KEY, FIRST_ADDRESS, FIRST_PRIVATE_KEY, OPERATOR_BLS_KEY,
         },
+        transaction::wait_transaction,
     };
     use hex::FromHex;
-    use std::str::FromStr;
-
-    async fn build_avs_registry_chain_reader(http_endpoint: String) -> AvsRegistryChainReader {
-        let registry_coordinator = get_registry_coordinator_address(http_endpoint.clone()).await;
-        let operator_state_retriever =
-            get_operator_state_retriever_address(http_endpoint.clone()).await;
-
-        AvsRegistryChainReader::new(
-            get_test_logger(),
-            registry_coordinator,
-            operator_state_retriever,
-            http_endpoint,
-        )
-        .await
-        .unwrap()
-    }
 
     #[tokio::test]
     async fn test_get_quorum_count() {
@@ -1235,6 +1282,53 @@ mod tests {
     }
 
     #[tokio::test]
+    async fn test_get_operator_restaked_strategies() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+        let avs_reader = build_avs_registry_chain_reader(http_endpoint.clone()).await;
+        let avs_writer =
+            build_avs_registry_chain_writer(http_endpoint.clone(), FIRST_PRIVATE_KEY.to_string())
+                .await;
+
+        let strategies = avs_reader
+            .get_operator_restaked_strategies(FIRST_ADDRESS)
+            .await
+            .unwrap();
+
+        assert!(strategies.is_empty());
+
+        let bls_key = OPERATOR_BLS_KEY.to_string();
+        let quorum_nums = Bytes::from([0]);
+        let bls_key_pair = BlsKeyPair::new(bls_key).unwrap();
+        let digest_hash: FixedBytes<32> = FixedBytes::from([0x02; 32]);
+        let signature_expiry = U256::MAX;
+
+        let tx_hash = avs_writer
+            .register_operator_in_quorum_with_avs_registry_coordinator(
+                bls_key_pair,
+                digest_hash,
+                signature_expiry,
+                quorum_nums,
+                "socket".to_string(),
+            )
+            .await
+            .unwrap();
+
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+
+        assert!(tx_status);
+
+        let strategies = avs_reader
+            .get_operator_restaked_strategies(FIRST_ADDRESS)
+            .await
+            .unwrap();
+
+        assert!(!strategies.is_empty());
+    }
+
+    #[tokio::test]
     async fn test_query_existing_registered_operator_sockets() {
         let (_container, http_endpoint, _ws_endpoint) = start_anvil_container().await;
         let avs_reader = build_avs_registry_chain_reader(http_endpoint.clone()).await;
@@ -1243,6 +1337,15 @@ mod tests {
             .query_existing_registered_operator_sockets(0, 1000)
             .await
             .unwrap();
+    }
+
+    #[tokio::test]
+    async fn test_get_restakeable_strategies() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+        let avs_reader = build_avs_registry_chain_reader(http_endpoint.clone()).await;
+
+        let strategies = avs_reader.get_restakeable_strategies().await.unwrap();
+        assert!(!strategies.is_empty());
     }
 
     #[tokio::test]