+++ crates/chainio/clients/avsregistry/src/writer.rs
@@ -1,27 +1,31 @@
 use crate::error::AvsRegistryError;
 use alloy::primitives::aliases::U96;
 use alloy::primitives::{Address, Bytes, FixedBytes, TxHash, U256};
+use alloy::providers::WalletProvider;
 use alloy::signers::local::PrivateKeySigner;
 use alloy::signers::Signer;
 use eigen_client_elcontracts::reader::ELChainReader;
+use eigen_common::{get_provider, get_signer};
 use eigen_crypto_bls::{
     alloy_g1_point_to_g1_affine, convert_to_g1_point, convert_to_g2_point, BlsKeyPair,
 };
 use eigen_logging::logger::SharedLogger;
+use eigen_types::operator::operator_id_from_g1_pub_key;
+use eigen_types::operator::QuorumNum;
+use eigen_utils::convert_stake_registry_strategy_params_to_registry_coordinator_strategy_params;
+use eigen_utils::slashing::middleware::registrycoordinator::IBLSApkRegistryTypes::PubkeyRegistrationParams;
+use eigen_utils::slashing::middleware::registrycoordinator::ISlashingRegistryCoordinatorTypes::OperatorKickParam;
 use eigen_utils::slashing::middleware::registrycoordinator::ISlashingRegistryCoordinatorTypes::OperatorSetParam;
-use eigen_utils::slashing::middleware::registrycoordinator::IStakeRegistryTypes::StrategyParams;
 use eigen_utils::slashing::middleware::registrycoordinator::{
-    IBLSApkRegistryTypes::PubkeyRegistrationParams, ISignatureUtils::SignatureWithSaltAndExpiry,
-    RegistryCoordinator,
+    ISignatureUtils::SignatureWithSaltAndExpiry, RegistryCoordinator,
 };
 use eigen_utils::slashing::middleware::servicemanagerbase::IRewardsCoordinatorTypes::OperatorDirectedRewardsSubmission;
+use eigen_utils::slashing::middleware::stakeregistry::IStakeRegistryTypes::StrategyParams;
 use eigen_utils::slashing::middleware::{
     servicemanagerbase::ServiceManagerBase, stakeregistry::StakeRegistry,
 };
-
-use eigen_common::{get_provider, get_signer};
 use std::str::FromStr;
-use tracing::info;
+use tracing::{info, warn};
 
 /// Gas limit for registerOperator in [`RegistryCoordinator`]
 pub const GAS_LIMIT_REGISTER_OPERATOR_REGISTRY_COORDINATOR: u64 = 2000000;
@@ -192,6 +196,154 @@ impl AvsRegistryChainWriter {
             .map_err(AvsRegistryError::AlloyContractError)?;
 
         info!(tx_hash = ?tx.tx_hash(),"Sent transaction to register operator in the AVS's registry coordinator" );
+        Ok(*tx.tx_hash())
+    }
+
+    /// Registers an operator while replacing existing operators in full quorums. If any quorum reaches
+    /// its maximum operator capacity, `operatorKickParams` is used to replace an old operator with the new one.
+    ///
+    /// # Arguments
+    ///
+    /// * `bls_key_pair` - bls key pair of the operator
+    /// * `operator_to_avs_registration_sig_salt` - operator signature salt
+    /// * `operator_to_avs_registration_sig_expiry` - operator signature expiry
+    /// * `quorum_numbers` - quorum numbers to register the new operator
+    /// * `socket` - socket used for calling the contract with `registerOperator` function
+    /// * `operators_to_kick` - operators to kick if quorum is full
+    /// * `churn_signer_private_key` - private key of the churn signer
+    /// * `churn_sig_salt` - churn signature salt
+    /// * `churn_sig_expiry` - churn signature expiry
+    ///
+    /// # Returns
+    ///
+    /// * `TxHash` - transaction hash of the register operator with churn transaction
+    #[allow(clippy::too_many_arguments)]
+    pub async fn register_operator_with_churn(
+        &self,
+        bls_key_pair: BlsKeyPair,
+        operator_to_avs_registration_sig_salt: FixedBytes<32>,
+        operator_to_avs_registration_sig_expiry: U256,
+        quorum_numbers: Bytes,
+        socket: String,
+        operators_to_kick: Vec<Address>,
+        churn_signer_private_key: String,
+        churn_sig_salt: FixedBytes<32>,
+        churn_sig_expiry: U256,
+    ) -> Result<TxHash, AvsRegistryError> {
+        let provider = get_signer(&self.signer.clone(), &self.provider);
+        let operator_wallet = PrivateKeySigner::from_str(&self.signer)
+            .map_err(|_| AvsRegistryError::InvalidPrivateKey)?;
+        let operator_address = operator_wallet.address();
+
+        info!(
+            avs_service_manager = %self.service_manager_addr,
+            operator = %operator_address,
+            quorum_numbers = ?quorum_numbers,
+            "registering operator with churn the AVS's registry coordinator"
+        );
+
+        let contract_registry_coordinator =
+            RegistryCoordinator::new(self.registry_coordinator_addr, &provider);
+
+        let g1_hashed_msg_to_sign = contract_registry_coordinator
+            .pubkeyRegistrationMessageHash(operator_address)
+            .call()
+            .await
+            .map_err(|_| AvsRegistryError::PubKeyRegistrationMessageHash)?
+            ._0;
+
+        let sig = bls_key_pair
+            .sign_hashed_to_curve_message(alloy_g1_point_to_g1_affine(g1_hashed_msg_to_sign))
+            .g1_point();
+        let alloy_g1_point_signed_msg = convert_to_g1_point(sig.g1())?;
+        let g1_pub_key_bn254 = convert_to_g1_point(bls_key_pair.public_key().g1())?;
+        let g2_pub_key_bn254 = convert_to_g2_point(bls_key_pair.public_key_g2().g2())?;
+
+        let pub_key_reg_params = PubkeyRegistrationParams {
+            pubkeyRegistrationSignature: alloy_g1_point_signed_msg.clone(),
+            pubkeyG1: g1_pub_key_bn254.clone(),
+            pubkeyG2: g2_pub_key_bn254.clone(),
+        };
+
+        let msg_to_sign = self
+            .el_reader
+            .calculate_operator_avs_registration_digest_hash(
+                operator_address,
+                self.service_manager_addr,
+                operator_to_avs_registration_sig_salt,
+                operator_to_avs_registration_sig_expiry,
+            )
+            .await?;
+
+        let operator_signature = operator_wallet
+            .sign_hash(&msg_to_sign)
+            .await
+            .map_err(|_| AvsRegistryError::InvalidSignature)?;
+
+        let operator_signature_with_salt_and_expiry = SignatureWithSaltAndExpiry {
+            signature: operator_signature.as_bytes().into(),
+            salt: operator_to_avs_registration_sig_salt,
+            expiry: operator_to_avs_registration_sig_expiry,
+        };
+
+        let operators_to_kick_params: Vec<OperatorKickParam> = operators_to_kick
+            .iter()
+            .zip(quorum_numbers.iter())
+            .map(|(address, quorum_number)| OperatorKickParam {
+                operator: *address,
+                quorumNumber: *quorum_number,
+            })
+            .collect();
+
+        let operator_id = FixedBytes::from(
+            operator_id_from_g1_pub_key(bls_key_pair.public_key())
+                .map_err(|_| AvsRegistryError::GetOperatorId)?,
+        );
+
+        let churn_wallet = PrivateKeySigner::from_str(&churn_signer_private_key)
+            .map_err(|_| AvsRegistryError::InvalidPrivateKey)?;
+
+        let churn_digest_hash = contract_registry_coordinator
+            .calculateOperatorChurnApprovalDigestHash(
+                operator_address,
+                operator_id,
+                operators_to_kick_params.clone(),
+                churn_sig_salt,
+                churn_sig_expiry,
+            )
+            .call()
+            .await?
+            ._0;
+
+        let churn_signature = churn_wallet
+            .sign_hash(&churn_digest_hash)
+            .await
+            .map_err(|_| AvsRegistryError::InvalidSignature)?;
+
+        let churn_signature_with_salt_and_expiry = SignatureWithSaltAndExpiry {
+            signature: churn_signature.as_bytes().into(),
+            salt: churn_sig_salt,
+            expiry: churn_sig_expiry,
+        };
+
+        let contract_call = contract_registry_coordinator.registerOperatorWithChurn(
+            quorum_numbers,
+            socket,
+            pub_key_reg_params,
+            operators_to_kick_params,
+            churn_signature_with_salt_and_expiry,
+            operator_signature_with_salt_and_expiry,
+        );
+
+        let tx = contract_call
+            .send()
+            .await
+            .map_err(AvsRegistryError::AlloyContractError)?;
+
+        info!(
+            tx_hash = ?tx.tx_hash(),
+            "Sent transaction to register operator with churn in the AVS's registry coordinator"
+        );
         Ok(*tx.tx_hash())
     }
 
@@ -497,7 +649,7 @@ impl AvsRegistryChainWriter {
     /// * `TxHash` - The transaction hash of the set operator set param transaction
     pub async fn set_operator_set_param(
         &self,
-        quorum_number: u8,
+        quorum_number: QuorumNum,
         operator_set_params: OperatorSetParam,
     ) -> Result<TxHash, AvsRegistryError> {
         info!("setting operator set param with the AVS's registry coordinator");
@@ -517,6 +669,176 @@ impl AvsRegistryChainWriter {
         Ok(*tx.tx_hash())
     }
 
+    /// Sets the minimum stake for the quorum
+    ///
+    /// Can only be called by the registry coordinator's owner.
+    /// The quorum number must exist, or else the tx will fail.
+    ///
+    /// # Arguments
+    ///
+    /// * `quorum_number` - The respective quorum number.
+    /// * `minimum_stake` - The minimum stake as a [`U96`].
+    ///
+    /// # Returns
+    ///
+    /// * [`TxHash`] - hash of the transaction.
+    pub async fn set_minimum_stake_for_quorum(
+        &self,
+        quorum_number: QuorumNum,
+        minimum_stake: U96,
+    ) -> Result<TxHash, AvsRegistryError> {
+        let provider = get_signer(&self.signer.clone(), &self.provider);
+
+        let contract_stake_registry = StakeRegistry::new(self.stake_registry_addr, provider);
+        let reg_coordinator_owner =
+            RegistryCoordinator::new(self.registry_coordinator_addr, get_provider(&self.provider))
+                .owner()
+                .call()
+                .await?
+                ._0;
+        let caller_address =
+            get_signer(&self.signer.clone(), &self.provider).default_signer_address();
+        if !reg_coordinator_owner.eq(&caller_address) {
+            warn!(caller = %caller_address,registry_coordinator_owner = %reg_coordinator_owner,"Caller must be registry coordinator's owner when calling set_minimum_stake_for_quorum");
+        }
+
+        let tx = contract_stake_registry
+            .setMinimumStakeForQuorum(quorum_number, minimum_stake)
+            .send()
+            .await
+            .map_err(AvsRegistryError::AlloyContractError)?;
+        info!(tx_hash = ?tx.tx_hash(),quorum_number = %quorum_number,"Setting minimum stake for quorum");
+        Ok(*tx.tx_hash())
+    }
+
+    /// Add strategies for the given quorum.
+    ///
+    /// Can only be called by the registry coordinator's owner.
+    /// The quorum numbers must exist, or else the tx will fail.
+    ///
+    /// # Arguments
+    ///
+    /// * `quorum_number` - The respective quorum's number.
+    /// * `strategy_params` - list of parameters for the strategies to add.
+    ///
+    /// # Returns
+    ///
+    /// * [`TxHash`] - hash of the transaction.
+    pub async fn add_strategies(
+        &self,
+        quorum_number: QuorumNum,
+        strategy_params: Vec<StrategyParams>,
+    ) -> Result<TxHash, AvsRegistryError> {
+        let provider = get_signer(&self.signer.clone(), &self.provider);
+
+        let contract_stake_registry = StakeRegistry::new(self.stake_registry_addr, provider);
+        let reg_coordinator_owner =
+            RegistryCoordinator::new(self.registry_coordinator_addr, get_provider(&self.provider))
+                .owner()
+                .call()
+                .await?
+                ._0;
+        let caller_address =
+            get_signer(&self.signer.clone(), &self.provider).default_signer_address();
+        if !reg_coordinator_owner.eq(&caller_address) {
+            warn!(caller = %caller_address,registry_coordinator_owner = %reg_coordinator_owner,"Caller must be registry coordinator's owner when adding strategies");
+        }
+        let tx = contract_stake_registry
+            .addStrategies(quorum_number, strategy_params)
+            .send()
+            .await
+            .map_err(AvsRegistryError::AlloyContractError)?;
+
+        info!(tx_hash = ?tx.tx_hash(),"Adding strategies");
+        Ok(*tx.tx_hash())
+    }
+
+    /// Remove strategies and their associated weights from the quorum's considered strategies.
+    ///
+    /// Can only be called by the registry coordinator's owner.
+    /// The quorum numbers must exist, else the tx will fail.
+    /// higher indices should be *first* in the list of `indices_to_remove`, since otherwise
+    /// the removal of lower index entries will cause a shift in the indices of the other strategies to remove
+    ///  
+    /// # Arguments
+    ///
+    /// # `quorum_number` - The respective quorum numbers in [`QuorumNum`].
+    /// # `indices_to_remove` - [`Vec<U256>`].
+    ///
+    /// # Returns
+    ///
+    /// * [`TxHash`] - The transaction hash of the transaction.
+    pub async fn remove_strategies(
+        &self,
+        quorum_number: QuorumNum,
+        indices_to_remove: Vec<U256>,
+    ) -> Result<TxHash, AvsRegistryError> {
+        let provider = get_signer(&self.signer.clone(), &self.provider);
+
+        let contract_stake_registry = StakeRegistry::new(self.stake_registry_addr, provider);
+        let reg_coordinator_owner =
+            RegistryCoordinator::new(self.registry_coordinator_addr, get_provider(&self.provider))
+                .owner()
+                .call()
+                .await?
+                ._0;
+        let caller_address =
+            get_signer(&self.signer.clone(), &self.provider).default_signer_address();
+        if !reg_coordinator_owner.eq(&caller_address) {
+            warn!(caller = %caller_address,registry_coordinator_owner = %reg_coordinator_owner,"Caller must be registry coordinator's owner when removing strategies");
+        }
+        let tx = contract_stake_registry
+            .removeStrategies(quorum_number, indices_to_remove)
+            .send()
+            .await
+            .map_err(AvsRegistryError::AlloyContractError)?;
+        info!(tx_hash = ?tx.tx_hash(),"Removing strategies");
+        Ok(*tx.tx_hash())
+    }
+
+    /// Modifies the weights of existing strategies for a specific quorum.
+    ///
+    /// Can only be called by the registry coordinator's owner.
+    /// The quorum numbers must exist, else the tx will fail.
+    ///  
+    /// # Arguments
+    ///
+    /// # `quorum_number` - The respective quorum numbers in [`QuorumNum`].
+    /// # `strategy_indices` - Indices of the strategies to change in [`Vec<U256>`].
+    /// # `new_multipliers` -  New multipliers for the strategies in [`Vec<U96>`].
+    ///
+    /// # Returns
+    ///
+    /// * [`TxHash`] - The transaction hash of the transaction.
+    pub async fn modify_strategy_params(
+        &self,
+        quorum_number: QuorumNum,
+        strategy_indices: Vec<U256>,
+        new_multipliers: Vec<U96>,
+    ) -> Result<TxHash, AvsRegistryError> {
+        let provider = get_signer(&self.signer.clone(), &self.provider);
+
+        let contract_stake_registry = StakeRegistry::new(self.stake_registry_addr, provider);
+        let reg_coordinator_owner =
+            RegistryCoordinator::new(self.registry_coordinator_addr, get_provider(&self.provider))
+                .owner()
+                .call()
+                .await?
+                ._0;
+        let caller_address =
+            get_signer(&self.signer.clone(), &self.provider).default_signer_address();
+        if !reg_coordinator_owner.eq(&caller_address) {
+            warn!(caller = %caller_address,registry_coordinator_owner = %reg_coordinator_owner,"Caller must be registry coordinator's owner when modifying strategy params");
+        }
+
+        let tx = contract_stake_registry
+            .modifyStrategyParams(quorum_number, strategy_indices, new_multipliers)
+            .send()
+            .await
+            .map_err(AvsRegistryError::AlloyContractError)?;
+        info!(tx_hash = ?tx.tx_hash(),"modifying strategy params");
+        Ok(*tx.tx_hash())
+    }
     /// Create a new quorum that tracks total delegated stake for operators.
     ///
     /// # Arguments
@@ -540,11 +862,18 @@ impl AvsRegistryChainWriter {
 
         let contract_registry_coordinator =
             RegistryCoordinator::new(self.registry_coordinator_addr, provider);
-
+        let registry_coordinator_strategy_params = strategy_params
+            .iter()
+            .map(|param| {
+                convert_stake_registry_strategy_params_to_registry_coordinator_strategy_params(
+                    param.clone(),
+                )
+            })
+            .collect();
         let contract_call = contract_registry_coordinator.createTotalDelegatedStakeQuorum(
             operator_set_param,
             minimum_stake,
-            strategy_params,
+            registry_coordinator_strategy_params,
         );
 
         contract_call
@@ -580,10 +909,18 @@ impl AvsRegistryChainWriter {
         let contract_registry_coordinator =
             RegistryCoordinator::new(self.registry_coordinator_addr, provider);
 
+        let registry_coordinator_strategy_params = strategy_params
+            .iter()
+            .map(|param| {
+                convert_stake_registry_strategy_params_to_registry_coordinator_strategy_params(
+                    param.clone(),
+                )
+            })
+            .collect();
         contract_registry_coordinator.createSlashableStakeQuorum(
             operator_set_param,
             minimum_stake,
-            strategy_params,
+            registry_coordinator_strategy_params,
             look_ahead_period,
             ).send()
             .await
@@ -651,16 +988,21 @@ mod tests {
         test_deregister_operator, test_register_operator,
     };
     use alloy::primitives::aliases::U96;
-    use alloy::primitives::U256;
-    use alloy::primitives::{Address, Bytes};
+    use alloy::primitives::{address, Address, Bytes, FixedBytes, U256};
     use alloy::sol_types::SolCall;
     use eigen_common::{get_provider, get_signer};
+    use eigen_crypto_bls::BlsKeyPair;
     use eigen_testing_utils::anvil::{start_anvil_container, start_m2_anvil_container};
-    use eigen_testing_utils::anvil_constants::get_allocation_manager_address;
     use eigen_testing_utils::anvil_constants::get_erc20_mock_strategy;
     use eigen_testing_utils::anvil_constants::get_rewards_coordinator_address;
     use eigen_testing_utils::anvil_constants::get_service_manager_address;
     use eigen_testing_utils::anvil_constants::GENESIS_TIMESTAMP;
+    use eigen_testing_utils::anvil_constants::SECOND_PRIVATE_KEY;
+    use eigen_testing_utils::anvil_constants::THIRD_ADDRESS;
+    use eigen_testing_utils::anvil_constants::THIRD_PRIVATE_KEY;
+    use eigen_testing_utils::anvil_constants::{
+        get_allocation_manager_address, OPERATOR_BLS_KEY_2,
+    };
     use eigen_testing_utils::anvil_constants::{
         FIFTH_ADDRESS, FIFTH_PRIVATE_KEY, FIRST_ADDRESS, FIRST_PRIVATE_KEY, OPERATOR_BLS_KEY,
         SECOND_ADDRESS,
@@ -669,12 +1011,12 @@ mod tests {
     use eigen_utils::slashing::core::allocationmanager::AllocationManager;
     use eigen_utils::slashing::core::irewardscoordinator::IRewardsCoordinator;
     use eigen_utils::slashing::middleware::registrycoordinator::ISlashingRegistryCoordinatorTypes::OperatorSetParam;
-    use eigen_utils::slashing::middleware::registrycoordinator::IStakeRegistryTypes::StrategyParams;
     use eigen_utils::slashing::middleware::registrycoordinator::RegistryCoordinator;
     use eigen_utils::slashing::middleware::servicemanagerbase::IRewardsCoordinatorTypes::OperatorDirectedRewardsSubmission;
     use eigen_utils::slashing::middleware::servicemanagerbase::IRewardsCoordinatorTypes::OperatorReward;
     use eigen_utils::slashing::middleware::servicemanagerbase::IRewardsCoordinatorTypes::StrategyAndMultiplier;
     use eigen_utils::slashing::middleware::servicemanagerbase::ServiceManagerBase;
+    use eigen_utils::slashing::middleware::stakeregistry::IStakeRegistryTypes::StrategyParams;
     use eigen_utils::slashing::middleware::stakeregistry::StakeRegistry;
     use futures_util::StreamExt;
 
@@ -1049,6 +1391,128 @@ mod tests {
     }
 
     #[tokio::test]
+    async fn test_register_operator_with_churn() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+        let bls_key = OPERATOR_BLS_KEY.to_string();
+        let private_key = FIRST_PRIVATE_KEY.to_string();
+        let quorum_nums = Bytes::from([0]);
+        let avs_writer =
+            build_avs_registry_chain_writer(http_endpoint.clone(), private_key.clone()).await;
+        let avs_reader = build_avs_registry_chain_reader(http_endpoint.clone()).await;
+
+        test_register_operator(
+            &avs_writer,
+            bls_key.clone(),
+            quorum_nums.clone(),
+            http_endpoint.clone(),
+        )
+        .await;
+
+        let is_registered = avs_reader
+            .is_operator_registered(FIRST_ADDRESS)
+            .await
+            .unwrap();
+        assert!(is_registered);
+
+        let operator_set_params = OperatorSetParam {
+            maxOperatorCount: 1,
+            kickBIPsOfOperatorStake: 10,
+            kickBIPsOfTotalStake: 10000,
+        };
+        let tx_hash = avs_writer
+            .set_operator_set_param(0, operator_set_params.clone())
+            .await
+            .unwrap();
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+        assert!(tx_status);
+
+        let tx_hash = avs_writer.set_churn_approver(THIRD_ADDRESS).await.unwrap();
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+        assert!(tx_status);
+
+        let avs_writer_2 =
+            build_avs_registry_chain_writer(http_endpoint.clone(), SECOND_PRIVATE_KEY.to_string())
+                .await;
+        let bls_key_2 = OPERATOR_BLS_KEY_2.to_string();
+
+        let operator_sig_salt = FixedBytes::from([0x02; 32]);
+        let churn_sig_salt = FixedBytes::from([0x05; 32]);
+        let sig_expiry = U256::MAX;
+        let churn_private_key = THIRD_PRIVATE_KEY.to_string();
+
+        let tx_hash = avs_writer_2
+            .register_operator_with_churn(
+                BlsKeyPair::new(bls_key_2).unwrap(),
+                operator_sig_salt,
+                sig_expiry,
+                quorum_nums.clone(),
+                "socket".to_string(),
+                vec![FIRST_ADDRESS],
+                churn_private_key,
+                churn_sig_salt,
+                sig_expiry,
+            )
+            .await
+            .unwrap();
+
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+        assert!(tx_status);
+
+        let avs_reader = build_avs_registry_chain_reader(http_endpoint.clone()).await;
+        let is_registered = avs_reader
+            .is_operator_registered(FIRST_ADDRESS)
+            .await
+            .unwrap();
+        assert!(!is_registered);
+
+        let avs_reader = build_avs_registry_chain_reader(http_endpoint.clone()).await;
+        let is_registered = avs_reader
+            .is_operator_registered(SECOND_ADDRESS)
+            .await
+            .unwrap();
+        assert!(is_registered);
+    }
+
+    #[tokio::test]
+    async fn test_set_minimum_stake_for_quorum() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+        let private_key = FIRST_PRIVATE_KEY.to_string();
+        let avs_writer = build_avs_registry_chain_writer(http_endpoint.clone(), private_key).await;
+        let quorum_number = 0;
+        let minimum_stake = U96::from(10);
+        let tx_hash = avs_writer
+            .set_minimum_stake_for_quorum(quorum_number, minimum_stake)
+            .await
+            .unwrap();
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+        assert!(tx_status);
+
+        let contract_stake_registry =
+            StakeRegistry::new(avs_writer.stake_registry_addr, get_provider(&http_endpoint));
+        assert_eq!(
+            contract_stake_registry
+                .minimumStakeForQuorum(quorum_number)
+                .call()
+                .await
+                .unwrap()
+                ._0,
+            minimum_stake
+        );
+    }
+
+    #[tokio::test]
     async fn test_create_total_delegated_stake_quorum() {
         let (_container, http_endpoint, _ws_endpoint) = start_anvil_container().await;
         let private_key = FIRST_PRIVATE_KEY.to_string();
@@ -1200,6 +1664,80 @@ mod tests {
     }
 
     #[tokio::test]
+    async fn test_set_add_strategies() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+
+        let private_key = FIRST_PRIVATE_KEY.to_string();
+
+        let avs_writer = build_avs_registry_chain_writer(http_endpoint.clone(), private_key).await;
+
+        let quorum_number = 0;
+        let strategy_params = [StrategyParams {
+            strategy: address!("54945180dB7943c0ed0FEE7EdaB2Bd24620256bc"),
+            multiplier: U96::from(1),
+        }];
+
+        let tx_hash = avs_writer
+            .add_strategies(quorum_number, strategy_params.to_vec())
+            .await
+            .unwrap();
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+
+        assert!(tx_status);
+
+        let contract_stake_registry =
+            StakeRegistry::new(avs_writer.stake_registry_addr, get_provider(&http_endpoint));
+        let expected_strategy = contract_stake_registry
+            .strategyParams(quorum_number, "1".parse().unwrap())
+            .call()
+            .await
+            .unwrap()
+            .strategy;
+
+        assert_eq!(strategy_params[0].strategy, expected_strategy);
+    }
+
+    #[tokio::test]
+    async fn test_remove_strategies() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+
+        let private_key = FIRST_PRIVATE_KEY.to_string();
+
+        let avs_writer = build_avs_registry_chain_writer(http_endpoint.clone(), private_key).await;
+
+        let quorum_number = 0;
+        let strategy_params = [StrategyParams {
+            strategy: address!("54945180dB7943c0ed0FEE7EdaB2Bd24620256bc"),
+            multiplier: U96::from(1),
+        }];
+
+        avs_writer
+            .add_strategies(quorum_number, strategy_params.to_vec())
+            .await
+            .unwrap();
+
+        let tx_hash = avs_writer
+            .remove_strategies(quorum_number, [U256::from(1)].to_vec())
+            .await
+            .unwrap();
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+
+        assert!(tx_status);
+        let contract_stake_registry =
+            StakeRegistry::new(avs_writer.stake_registry_addr, get_provider(&http_endpoint));
+        assert!(contract_stake_registry
+            .strategyParams(quorum_number, "1".parse().unwrap())
+            .call()
+            .await
+            .is_err());
+    }
+    #[tokio::test]
     async fn test_set_ejector_cooldown() {
         let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
         let private_key = FIRST_PRIVATE_KEY.to_string();
@@ -1263,14 +1801,56 @@ mod tests {
             .await
             .unwrap()
             .status();
-
         assert!(tx_status);
+    }
 
-        let is_registered = avs_reader
-            .is_operator_registered(register_operator_address)
+    #[tokio::test]
+    async fn test_modify_strategy_params() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+
+        let private_key = FIRST_PRIVATE_KEY.to_string();
+
+        let avs_writer = build_avs_registry_chain_writer(http_endpoint.clone(), private_key).await;
+
+        let quorum_number = 0;
+        let strategy_params = [StrategyParams {
+            strategy: address!("54945180dB7943c0ed0FEE7EdaB2Bd24620256bc"),
+            multiplier: U96::from(1),
+        }];
+
+        avs_writer
+            .add_strategies(quorum_number, strategy_params.to_vec())
             .await
             .unwrap();
-        assert!(!is_registered);
+
+        let strategy_indices = [U256::from(1)];
+        let new_multipliers = [U96::from(2)];
+
+        let tx_hash = avs_writer
+            .modify_strategy_params(
+                quorum_number,
+                strategy_indices.to_vec(),
+                new_multipliers.to_vec(),
+            )
+            .await
+            .unwrap();
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+        assert!(tx_status);
+
+        let contract_stake_registry =
+            StakeRegistry::new(avs_writer.stake_registry_addr, get_provider(&http_endpoint));
+        assert_eq!(
+            U96::from(2),
+            contract_stake_registry
+                .strategyParams(quorum_number, "1".parse().unwrap())
+                .call()
+                .await
+                .unwrap()
+                .multiplier
+        );
     }
 
     #[tokio::test]